Structural refactoring to allow for easier log configuration
The client is moved into a submodule, so that it's possible to provide
a variety of default log configurations before importing the
module (which still triggers logging config automatically).

The goal is to have default usage Just Work, but allow for
configurability up to the limit of what structlog and logging allow.