Radical simplification of JSONModel stuff.
A relation is now pretty much a route and the params used, which can
be iterated over for all records, or can be used to search for an
individual record.

JSONModel object now essentially try to do an API call if
they don't have a matching key in their __json - this seems to pretty
much handle repository-scoping without special casing or obvious
breakage.  It's actually surprising how much of the "read"
functionality of the API this covers. I don't doubt we'll need some
special casing or want to collapse some routes, and we'll definitely
need to handle trees specially, for example, but the abstraction is
able to cover most objects index and show methods.