Change "authorize" POST to send password in request body
Currently, the "authorize" method in "asnake/client/web_client.py"
issues a POST request that sends the user's password via "params",
which means it ends up in the query string of the URL, for example:

```
https://aspace.example.org/users/jsmith/login?password=secret_password&expiring=False
```

This is problematic, because query string parameters are often visible
in web server log files.

This pull request simply switches "params" to "data", so that the
password is sent in the POST request body, and so will not show in the
web server log.

**NOTE:** I was not able to quickly figure out how to run the project's
unit/integration tests, so apologies in advance if they are broken
because of this change.